pyproject.toml
[project]
name = "levi"
version = "0.2.0"
description = "Voice-Driven AI Assistant for Spanish/English Translation"
requires-python = ">=3.13"

dependencies = [
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
levi = "levi_cli.main:cli"

[project.optional-dependencies]
mac = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mlx-lm>=0.19.0",
    "mlx-audio>=0.2.0",
    "websockets>=12.0",
    "transformers>=5.0.0",
    "torch>=2.10.0",
    "sentencepiece>=0.2.1",
    "protobuf>=6.33.4",
]

cloud = [
    "python-telegram-bot>=21.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "aiohttp>=3.9.0",
    "ntgcalls>=1.0.0",
    "pyrogram>=2.0.0",
    "TgCrypto>=1.2.5",
    "webrtcvad>=2.0.10",
]

cli = [
    "click>=8.1.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["levi_cli", "mac", "cloud"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
